renamed sc_test to sample_template. created create_entry and update_entry methods along with init db for sample template.
